:construction: Initialize CTCoreNet and CTCoreData modules
So it begins, with a bunch of data preprocessing and neural network code. Using Pytorch Lightning makes things very structured, though there is still a bit of boilerplate. Since this is a custom semantic segmentation task, most of the time is spent on figuring out the data preprocessing stuff and generating data labels for the images. Will follow up with a better model architecture and upload some sample data in the coming days.